split update version
