+++ .github/workflows/update_versions.yml
@@ -1,0 +1,124 @@
+# Workflow to handle packaging the Unity SDK
+name: Update Versions
+
+on:
+  workflow_dispatch:
+    inputs:
+      operating_systems:
+        description: 'CSV of VMs to run on'
+        default: 'ubuntu-latest'
+        required: true
+      package_version_number:
+        description: "The Unity SDK version to upgrade to"
+        default: 8.6.0
+
+jobs:
+  update_versions:
+    name: package-${{matrix.os}}
+    runs-on: ${{matrix.os}}
+    strategy:
+      fail-fast: false
+      matrix:
+        os: [ubuntu-latest]
+        include:
+        - os: ubuntu-latest
+          package_dir: output
+    steps:
+      - name: Check out base branch
+        uses: actions/checkout@v2.3.1
+        with:
+          path: firebase-unity-sdk
+          fetch-depth: 0
+          ref: ${{ github.event.inputs.baseBranch }}
+      
+      - name: Checkout CPP Repo
+        uses: actions/checkout@v2
+        with:
+          repository: firebase/firebase-cpp-sdk
+          path: firebase-cpp-sdk
+          ref: ${{ github.event.inputs.firebase_cpp_sdk_version }}
+          submodules: true
+      
+      - name: Get token for firebase-workflow-trigger
+        uses: tibdex/github-app-token@v1
+        id: generate-token
+        with:
+          app_id: ${{ secrets.WORKFLOW_TRIGGER_APP_ID }}
+          private_key: ${{ secrets.WORKFLOW_TRIGGER_APP_PRIVATE_KEY }}
+
+      - name: Setup Python
+        uses: actions/setup-python@v2
+        with:
+          python-version: 3.7
+
+      - name: Install python deps
+        shell: bash
+        run: |
+          cd firebase-unity-sdk
+          pip install -r scripts/gha/requirements.txt
+
+      - name: Name new branch
+        run: |
+          date_str=$(date "+%Y%m%d-%H%M%S")
+          echo "NEW_BRANCH=${{env.branchPrefix}}${{github.event.inputs.package_version_number}}-${date_str}" >> $GITHUB_ENV
+
+      - name: Create new branch
+        run: |
+          cd firebase-unity-sdk
+          git remote update
+          git checkout -b "${NEW_BRANCH}"
+          echo "UPDATE_LOGFILE=update_log.txt" >> $GITHUB_ENV
+
+      - name: Update Unity SDK version and dependencies
+        run: |
+          cd firebase-unity-sdk
+          python scripts/update_versions.py --unity_sdk_version=${{github.event.inputs.package_version_number}}
+
+  trigger_build_sdks:
+    needs: [update_versions]
+    name: trigger-build-${{matrix.platform}}
+    runs-on: ubuntu-latest
+    strategy:
+      fail-fast: false
+      matrix:
+        platform: [android, ios, desktop]
+        include:
+        - platform: android
+          trigger_file: android.yml
+        - platform: ios
+          trigger_file: ios.yml
+        - platform: desktop
+          trigger_file: sdk_build.yml
+
+    steps:
+      - name: Checkout repo
+        uses: actions/checkout@v2.3.1
+
+      - name: Setup python
+        uses: actions/setup-python@v2
+        with:
+          python-version: 3.7
+
+      - name: Generate token for GitHub API
+        # This step is necessary because the existing GITHUB_TOKEN cannot be used inside one workflow to trigger another.
+        # 
+        # Instead, generate a new token here, using our GitHub App's private key and App ID (saved as Secrets).
+        # 
+        # This method is preferred over the "personal access token" solution, as the GitHub App's scope is limited to just
+        # the firebase-cpp-sdk repository.
+        uses: tibdex/github-app-token@v1
+        id: generate-token
+        with:
+          app_id: ${{ secrets.WORKFLOW_TRIGGER_APP_ID }}
+          private_key: ${{ secrets.WORKFLOW_TRIGGER_APP_PRIVATE_KEY }}
+
+      - name: Use GitHub API to start workflow
+        shell: bash
+        run: |
+          pip install -r scripts/gha/requirements.txt
+          if [[ "${{ github.event_name }}" == "schedule" ]]; then
+            # reuse flag --test_pull_request=nightly-packaging to generate report
+            generate_report=(-p test_pull_request nightly-packaging)
+          fi
+          set -e
+          python scripts/gha/trigger_workflow.py -t ${{ steps.generate-token.outputs.token }} -w ${{ matrix.trigger_file }} -p test_packaged_sdk ${{ github.run_id }} -s 10 -A -v
