+++ cmake/firebase_unity_version.cmake
@@ -14,10 +14,10 @@
 
 # This file defines the version numbers used by the Firebase Unity SDK.
 
-set(FIREBASE_UNITY_SDK_VERSION "8.5.0"
+set(FIREBASE_UNITY_SDK_VERSION "8.6.0"
     CACHE STRING "The version of the Unity SDK, used in the names of files.")
 
-set(FIREBASE_IOS_POD_VERSION "8.3.0"
+set(FIREBASE_IOS_POD_VERSION "8.10.0"
     CACHE STRING "The version of the top-level Firebase Cocoapod to use.")
 
 # https://github.com/googlesamples/unity-jar-resolver
